rpc: stabilize the list and subscription APIs into sui.rpc.v2
Update the sui-rust-sdk dependency to rev 05018477, which moves the
filtered List and Subscribe APIs from the experimental sui.rpc.v2alpha
proto package into the stable sui.rpc.v2 package; only the proof APIs
remain in v2alpha.

- sui-rpc-api: move the list handlers from grpc/v2alpha/ledger_service
  into grpc/v2/ledger_service and merge the ListCheckpoints,
  ListTransactions, and ListEvents RPCs into the v2 LedgerService impl.
  Promote the v2alpha subscription implementation to the v2
  SubscriptionService and drop the subscribe_checkpoints_stable shim;
  SubscribeCheckpoints now validates read masks and honors the new
  `filter` field. Drop the separate v2alpha ledger and subscription
  service registrations; the proof service stays in v2alpha.
- sui-kv-rpc: apply the same merge to the BigTable-backed list handlers.
  Since the flag-gated v2alpha service no longer exists as a separate
  registration, `enable_experimental_query_apis` now gates the list RPCs
  at runtime: when disabled they return Unimplemented, matching the
  status clients previously received from the unregistered service.
- sui-light-client: drive the list APIs through the single v2
  LedgerServiceClient; the proof-service usage stays on v2alpha.
- Switch sui-indexer-alt-reader, sui-config, and the e2e test suites to
  the v2 types, including the QueryEndReason zero-value rename from
  UNSPECIFIED to UNKNOWN.
