+++ crates/sui-kv-rpc/src/v2/list_checkpoints.rs
@@ -18,11 +18,11 @@ use sui_rpc::field::FieldMask;
 use sui_rpc::field::FieldMaskTree;
 use sui_rpc::field::FieldMaskUtil;
 use sui_rpc::proto::sui::rpc::v2::Checkpoint;
-use sui_rpc::proto::sui::rpc::v2alpha::ListCheckpointsRequest;
-use sui_rpc::proto::sui::rpc::v2alpha::ListCheckpointsResponse;
-use sui_rpc::proto::sui::rpc::v2alpha::QueryEnd;
-use sui_rpc::proto::sui::rpc::v2alpha::QueryEndReason;
-use sui_rpc::proto::sui::rpc::v2alpha::Watermark;
+use sui_rpc::proto::sui::rpc::v2::ListCheckpointsRequest;
+use sui_rpc::proto::sui::rpc::v2::ListCheckpointsResponse;
+use sui_rpc::proto::sui::rpc::v2::QueryEnd;
+use sui_rpc::proto::sui::rpc::v2::QueryEndReason;
+use sui_rpc::proto::sui::rpc::v2::Watermark;
 use sui_rpc_api::ErrorReason;
 use sui_rpc_api::RpcError;
 use sui_rpc_api::ledger_history::query_options::CheckpointRange;
@@ -598,8 +598,8 @@ mod tests {
     use super::*;
     use sui_rpc_cursor::CursorToken;
 
-    use crate::v2alpha::test_utils::ascending_options;
-    use crate::v2alpha::test_utils::query_context;
+    use crate::v2::test_utils::ascending_options;
+    use crate::v2::test_utils::query_context;
 
     #[tokio::test]
     async fn natural_empty_range_emits_one_standalone_terminal_boundary() {
@@ -747,7 +747,7 @@ mod tests {
             let mut proto_options = ascending_options();
             if !direction.is_ascending() {
                 proto_options.ordering =
-                    Some(sui_rpc::proto::sui::rpc::v2alpha::Ordering::Descending as i32);
+                    Some(sui_rpc::proto::sui::rpc::v2::Ordering::Descending as i32);
             }
             let options =
                 QueryOptions::checkpoints_from_proto(Some(&proto_options), 10, 100).unwrap();
@@ -796,7 +796,7 @@ mod tests {
             let mut proto_options = ascending_options();
             if !direction.is_ascending() {
                 proto_options.ordering =
-                    Some(sui_rpc::proto::sui::rpc::v2alpha::Ordering::Descending as i32);
+                    Some(sui_rpc::proto::sui::rpc::v2::Ordering::Descending as i32);
             }
             let options =
                 QueryOptions::checkpoints_from_proto(Some(&proto_options), 10, 100).unwrap();
