+++ crates/sui-rpc-api/src/grpc/v2/ledger_service/chunked_scan.rs
@@ -3,7 +3,7 @@
 
 use std::collections::VecDeque;
 
-use sui_rpc::proto::sui::rpc::v2alpha::Watermark;
+use sui_rpc::proto::sui::rpc::v2::Watermark;
 use sui_rpc_cursor::Position;
 use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
