+++ crates/sui-rpc-api/src/grpc/v2/ledger_service/query_end.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use sui_rpc::proto::sui::rpc::v2alpha::QueryEndReason;
+use sui_rpc::proto::sui::rpc::v2::QueryEndReason;
 
 /// Effective terminal reason for a successful query stream.
 ///
