+++ crates/sui-rpc-api/src/grpc/v2alpha/mod.rs
@@ -1,6 +1,4 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-mod ledger_service;
 mod proof_service;
-pub(crate) mod subscription_service;
